+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [3.2.0] - 2026-07-29
 
 ### Added
 
@@ -328,7 +328,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release of standalone version
 
 
-[unreleased]: https://github.com/fkleon/math-expressions/compare/3.1.0...HEAD
+[unreleased]: https://github.com/fkleon/math-expressions/compare/3.2.0...HEAD
+[3.2.0]: https://github.com/fkleon/math-expressions/compare/3.1.0...3.2.0
 [3.1.0]: https://github.com/fkleon/math-expressions/compare/3.0.0...3.1.0
 [3.0.0]: https://github.com/fkleon/math-expressions/compare/2.7.0...3.0.0
 [2.7.0]: https://github.com/fkleon/math-expressions/compare/2.6.0...2.7.0